+++ examples/nextjs-cookie/next.config.mjs
@@ -3,6 +3,7 @@ import { withPalamedes } from "@palamedes/next-plugin"
 export default withPalamedes(
   {},
   {
+    messageSplitting: true,
     serverFunctions: true,
   }
 )